Modified balance double value to BigDecimal

+++ src/main/java/com/paulo/minipicpay/model/PicPay.java
@@ -1,5 +1,7 @@
 package com.paulo.minipicpay.model;
 
+import java.math.BigDecimal;
+
 public class PicPay {
 
     private final UserRepository userRepository;
@@ -10,14 +12,14 @@ public class PicPay {
         this.validateHandler = validateHandler;
     }
 
-    public void transfer(User sender, User recipient, Double amount) {
+    public void transfer(User sender, User recipient, BigDecimal amount) {
         validateHandler.validate(sender, amount);
         sender.decreaseBalance(amount);
         recipient.increaseBalance(amount);
         updateUsersBalance(sender, recipient, amount);
     }
 
-    private void updateUsersBalance(User sender, User recipient, Double amount) {
+    private void updateUsersBalance(User sender, User recipient, BigDecimal amount) {
         userRepository.updateBalance(sender, amount);
         userRepository.updateBalance(recipient, amount);
     }
